Move DeleteForm to where it is used
